+++ gha_utils/metadata.py
@@ -377,6 +377,7 @@ class TargetVersion(StrEnum):
     PY311 = "3.11"
     PY312 = "3.12"
     PY313 = "3.13"
+    PY314 = "3.14"
 
 
 MYPY_VERSION_MIN: Final = (3, 8)