+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [4.19.3 (unreleased)](https://github.com/kdeldycke/workflows/compare/v4.19.2...main)
+
+> [!IMPORTANT]
+> This version is not released yet and is under active development.
+
 ## [4.19.2 (2025-09-26)](https://github.com/kdeldycke/workflows/compare/v4.19.1...v4.19.2)
 
 - Add new fields to `gha-utils metadata`: